moved all `.js` files into `src/` & removed TLog socket plugin
