+++ gfycat/adjectives.txt
