+++ src/generators/gfycat.js
@@ -1,6 +1,4 @@
 const fs = require('fs-extra');
-const adjectives = fs.readFileSync('./generators/gfycat/adjectives.txt').toString().split('\n');
-const animals = fs.readFileSync('./generators/gfycat/animals.txt').toString().split('\n');
 
 // Don't trigger circular dependency during setup
 if (!require.main.filename.includes('setup.js'))
@@ -11,6 +9,11 @@ function getWord(list, delim = '') {
 }
 
 function genString(count = MIN_LENGTH) {
+    // For some reason these 3 lines MUST be inside the function
+    const { path } = require('../utils');
+    const adjectives = fs.readFileSync(path('./gfycat/adjectives.txt')).toString().split('\n');
+    const animals = fs.readFileSync(path('./gfycat/animals.txt')).toString().split('\n');
+
     let gfycat = '';
     for (let i = 0; i < (count < MIN_LENGTH ? MIN_LENGTH : count); i++)
         gfycat += getWord(adjectives, '-');
