+++ src/generators/random.js
