+++ src/hash.js
