+++ src/setup.js
@@ -53,7 +53,7 @@ function getConfirmSchema(description) {
 
 // If directly called on the command line, run setup script
 function doSetup() {
-	const path = (...paths) => require('path').join(__dirname, ...paths);
+	const path = (...paths) => require('path').join(process.cwd(), ...paths);
 	const TLog = require('@tycrek/log');
 	const fs = require('fs-extra');
 	const prompt = require('prompt');
@@ -64,7 +64,7 @@ function doSetup() {
 	// Override default configs with existing configs to allow migrating configs
 	// Now that's a lot of configs!
 	try {
-		const existingConfig = require('./config.json');
+		const existingConfig = require('../config.json');
 		Object.entries(existingConfig).forEach(([key, value]) => {
 			Object.prototype.hasOwnProperty.call(config, key) && (config[key] = value); // skipcq: JS-0093
 			Object.prototype.hasOwnProperty.call(s3config, key) && (s3config[key] = value); // skipcq: JS-0093
