+++ src/vibrant.js
